requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
langgraph>=0.0.20

# Vector stores and embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

# Evaluation metrics
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13

# Data handling
pandas>=2.0.0
datasets>=2.16.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# API clients
groq>=0.4.0
